requirements.txt
numpy==1.26.3
pandas==2.0.3
tensorflow==2.16.1
keras==3.3.3
scikit-learn==1.3.0
Pillow==10.0.0
streamlit==1.30.0

pydantic
langchain-ollama
langchain-core
langchain-community
